// src/model/GSI.Model.ts
export namespace GSI {


export interface GameState {
  abilities?: HeroAbilities;
  auth?: Auth;
  hero?: Hero;
  items?: PlayerItems;
  map?: Map;
  player?: Player;
  previously?: PreviousGameState;
  provider?: Provider;
  draft?: {
    [k: string]: any;
  };
  wearables?: Wearables;
  buildings?: Buildings;
}

export interface HeroAbilities {
  ability0?: AbilityBase;
  ability1?: AbilityBase;
  ability2?: AbilityBase;
  ability3?: AbilityBase;
}

export interface AbilityBase {
  ability_active?: boolean;
  can_cast?: boolean;
  cooldown?: number;
  level?: number;
  name: E.NpcDotaHeroAbility;
  passive?: boolean;
  ultimate?: boolean;
}

export interface Auth {
  token?: string;
}

export interface Hero {
  alive?: boolean;
  break?: boolean;
  buyback_cooldown?: number;
  buyback_cost?: number;
  disarmed?: boolean;
  has_debuff?: boolean;
  health?: number;
  health_percent?: number;
  hexed?: boolean;
  id?: number;
  level?: number;
  magicimmune?: boolean;
  mana?: number;
  mana_percent?: number;
  max_health?: number;
  max_mana?: number;
  muted?: boolean;
  name?: E.NpcDotaHero;
  respawn_seconds?: number;
  silenced?: boolean;
  stunned?: boolean;
  xpos?: number;
  ypos?: number;
}

export interface PlayerItems {
  slot0?: ItemBase;
  slot1?: ItemBase;
  slot2?: ItemBase;
  slot3?: ItemBase;
  slot4?: ItemBase;
  slot5?: ItemBase;
  slot6?: ItemBase;
  slot7?: ItemBase;
  slot8?: ItemBase;
  stash0?: ItemBase;
  stash1?: ItemBase;
  stash2?: ItemBase;
  stash3?: ItemBase;
  stash4?: ItemBase;
  stash5?: ItemBase;
}

export interface ItemBase {
  can_cast?: boolean;
  charges?: number;
  contains_rune?: boolean;
  cooldown?: number;
  name: E.NpcDotaItems;
  passive?: boolean;
  purchaser?: number;
  [k: string]: any;
}
/**
 * GSI..Map
 */
export interface Map {
  clock_time?: number;
  customgamename?: string;
  daytime?: boolean;
  game_state?: E.GameState;
  game_time?: number;
  matchid?: number;
  name?: string;
  nightstalker_night?: boolean;
  paused?: boolean;
  ward_purchase_cooldown?: number;
  win_team?: E.PlayerTeam;
}

export interface Player {
  activity?: E.PlayerActivity;
  assists?: number;
  deaths?: number;
  denies?: number;
  gold?: number;
  gold_reliable?: number;
  gold_unreliable?: number;
  gpm?: number;
  kill_list?: {
    [k: string]: any;
  };
  kill_streak?: number;
  kills?: number;
  last_hits?: number;
  name?: string;
  pro_name?: string;
  steamid?: string;
  team_name?: E.PlayerTeam;
  xpm?: number;
}

export interface PreviousGameState {
  abilities?: boolean | HeroAbilities;
  auth?: boolean | Auth;
  hero?: boolean | Hero;
  items?: boolean | PlayerItems;
  map?: boolean | Map;
  player?: boolean | Player;
  provider?: boolean | Provider;
  draft?: boolean | { [k: string]: any;};
  wearables?: boolean | Wearables;
  buildings?: boolean | Buildings;
}
/**
 * Провайдер предоставляющий даные.
 */
export interface Provider {
  appid: E.AppSteamId;
  name: E.AppName;
  timestamp: number;
  /**
   * Версия протокола клиента.
   */
  version: 46;
}

export interface Wearables {
  [k: string]: any;
}

export interface Buildings {
  radiant?: RadiantBuilding;
}
// ...
}